clean-up UX code and recording this as an Alternative UX Model
