+++ ui-js/src/components/Layout/Layout.js
@@ -13,8 +13,8 @@ import ClinicInvite from "../../pages/clinicinvite/Clinicinvite";
 import CitizenInvite from "../../pages/citizeninvite/Citizeninvite";
 import TestRequest from "../../pages/testrequest/Testrequest";
 import TestAppointment from "../../pages/testappointment/Testappointment";
-import Covid19Test from "../../pages/covid19test/Covid19test";
-import StartForm from "../../pages/finalform/Start";
+import Covid19Test from "../../pages/covid19test/Covid19test_alt";
+import StartForm from "../../pages/finalform/Start_alt";
 import { wsBaseUrl, httpBaseUrl } from "../../config";
 
 
@@ -42,7 +42,7 @@ function Layout() {
                 <Route path="/app/citizeninvite" component={CitizenInvite} />
                 <Route path="/app/testrequest" component={TestRequest} />
                 <Route path="/app/testappointment" component={TestAppointment} />
-                <Route path="/app/covid19test" component={Covid19Test} />
+                <Route path="/app/covid19testalt" component={Covid19Test} />
                 <Route path="/app/finalform" component={StartForm} />
               </Switch>
             </div>
